Worker::GetString: fix segfault
There's a very hard to debug race condition around here. This is not
a proper fix, but with this, we're only "losing" frames instead of
plain segfaulting the tracy profiler.

I've been investigating this bug for a while, something very weird
happens on the client side. As soon as I send the column/line position
in the trace, some "???" frames appear. Hovering those frames
segfaults tracy.

I've been running this in GDB on both sides (client & profiler), and
as soon as we slow down the client, the "???" frames disappear. This
seem to point towards a race condition.

Instead of spending more hours understanding the client internals, I'm
rather going to go for this temporary fix and test this profiler on
Nixpkgs, figuring out if it worth investing more time fixing this race
condition or not.